Correct spelling in comments. [ci skip]
The change in src/Relude/Extra/Lens.h was a change of an identifier name in a comment. Specifically: addressCityL had an extra 'd'

In src/Relude/Extra/CallStack.hs line 34, the spelling of an "<unknown>" in a comment was wrong. The spelling in the code was correct and was not changed.

In src/Relude/Nub.hs, I corrected the spelling to "recommendations", and "results". Both were on the same line.

I had to change "functions" from plural to singular in src/Relude/Extra/Newtype.hs (line 93)
I changed "if" to "is" to make a sentence grammatical in src/Relude/Extra/Validation.hs (line 167)

The rest of the changes are much simpler, and straightforward.