Doc: Extra.Foldable1: foldr1: diff to base Data.Foldable [ci skip]
While migrating pretty big codebase (HNix) `foldr1` replacement is the only
place I tripped over and was puzzled a bit, the solution was simple, but harder
to see because difference was unexpected.

Since `relude` is very transparent in migration. It is good to document where
the API differences is, to guide people during migration and during use.

For more info:
https://github.com/kowainik/relude/issues/373
https://github.com/kowainik/relude/issues/307#issuecomment-798973946