README: add note on how to do mixins in the stack/hpack [ci skip]
It was hard to find the info how to do mixins in hpack or how to do relude with
hpack or hpack&stack.

Then I strumbled on the Reddit suggestion
https://www.reddit.com/r/haskelltil/comments/9qa366/easy_way_to_replace_default_prelude_with_the/
lead to result.

After that I noted that there is indeed mixins mention & examples in the `hpack`
ReadMe, I guess I not looked good enough in there, but those examples need to be
pieced together & their combination of use needs to be guessed a bit to
configure hpack for `relude`. So people would not guess how to use project with
`stack` or `hpack`, adding info of a
working setup additional example how to do versioning in this case.